update e2e tests
now takes 10% less time with use of cypress session
